src: Adding source files
ops.cpp and ops.hpp keep most of the compute API and wrapper functions
together.

simple.cpp and simple.hpp include all of the remaining code that: reads
test data from the sample file, peeks at the test data, re-organizes (if
necessary) the test data so that the compute functions can be applied
and timed appropriately.