simple.cpp: better control of execution params
This commit includes 3 main additions: (1) some print statements wrapped
in macros to check for DEBUG mode, (2) basic handling of additional CLI
args for processing multiple slices at a time or limiting the number of
columns processed, and (3) implementation of reading multiple slices and
concatenating them for processing or setting a stop column index so that
a specified number of columns are processed.

The 3rd point allows us to take a single dataset (the test dataset) and
better investigate how performance changes for a varying number of rows
(in multiples of chunk size) and for a varying number of columns (but as
a sub-range of the total set of columns).

Note: I tried adding code for breaking the dataset up into evenly sized
vertical partitions (and I left the code in, but commented out). It
turns out, unsurprisingly, that the time complexity of the algorithm
isn't sensitive at all to iterative applications over subsets of the
columns.